+++ CHANGELOG.md
@@ -43,6 +43,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
   bsn_activation_height and finality_signature_interval for spam protection
 * [97](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/97) feat:
   versioning of FP allowlist
+* [100](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/100) chore:
+  add query to fetch highest voted height
 
 ### Improvements